change indentSize in CodeMirrorEditor
